requirements.txt
beautifulsoup4==4.10.0
cryptography==36.0.2
gurux-dlms==1.0.121
lxml==4.8.0
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==0.19.2
